Prawie Gotowe

+++ funkcje.php
@@ -31,12 +31,29 @@ function addclient()
 		$bill = $_POST['bill'];
 
 		global $pdo;
-		$stmt = $pdo->query("INSERT INTO klienci SET Imie = '$firstname', Nazwisko = '$lastname', Samochod = '$car', Kontakt = '$contact', Rachunek 	='$bill'");	$stmt->closeCursor();
+		$stmt = $pdo->query("INSERT INTO klienci SET Imie = '$firstname', Nazwisko = '$lastname', Samochod = '$car', Kontakt = '$contact', Rachunek 	='$bill'");	
+		$stmt->closeCursor();
 		$stmt=null;
-	
-		header('Location: index.php');	
+		addcar();
+		echo '<div class="noerror">Klient został dodany do bazy.</div>';	
 	}
 	
+}
+
+function addcar()
+{
+	$firstname = $_POST['firstname'];
+	$lastname = $_POST['lastname'];
+	global $pdo;
+	$stmt = $pdo->query("SELECT ID, Samochod FROM klienci WHERE Imie = '$firstname' AND Nazwisko = '$lastname'");
+	foreach($stmt as $row)
+	{
+		$id = $row['ID'];
+		$car = $row['Samochod'];
+	}
+	$stmt = $pdo->query("INSERT INTO cars SET CAR = '$car', REP = 0, DONE = 0, klienci_id = '$id'");	
+	$stmt->closeCursor();
+	$stmt=null;
 }
 
 function editclient()
@@ -59,22 +76,66 @@ function editclient()
 		$stmt->closeCursor();
 		$stmt=null;
 
-		header('Location: index.php');	
+		echo '<div class="noerror">Dane klienta zostały zmienione.</div>';
 	}
 	
 }
 
 function deleteclient()
 {
-	$firstname = $_POST['firstname'];
-	$lastname = $_POST['lastname'];
-	$id = $_POST['id'];
+	if(empty($_POST['id']) || empty($_POST['firstname']) || empty($_POST['lastname']))
+	{
+		echo '<div class="error">Wystąpił błąd podczas usuwania klienta! Upewnij się, że wypełniłeś wszystkie pola formularza.</div>';	
+	}
+	else
+	{
+		$firstname = $_POST['firstname'];
+		$lastname = $_POST['lastname'];
+		$id = $_POST['id'];
 
-	$stmt = $pdo->query("DELETE FROM klienci WHERE ID ='$id'");			
-	$stmt->closeCursor();
-	$stmt=null;	
+		global $pdo;
+		$stmt = $pdo->query("DELETE FROM klienci WHERE ID ='$id'");			
+		$stmt->closeCursor();
+		$stmt=null;	
+		
+		echo '<div class="noerror">Klient został usunięty z bazy.</div>';
+	}
+}
+function repair()
+{
+	if(isset($_POST['repair']))
+	{
+	 	$id = $_GET['id'];
+		global $pdo;
+		$stmt = $pdo->query("UPDATE cars SET DONE = 1 WHERE ID = '$id'");			
+		$stmt->closeCursor();
+		$stmt=null;  
+	}
+	else
+	{
+		$id = $_GET['id'];
+		global $pdo;
+		$stmt = $pdo->query("UPDATE cars SET DONE = 0 WHERE ID = '$id'");			
+		$stmt->closeCursor();
+		$stmt=null;     
+	}
 	
-	header('Location: index.php');
+}
+
+function adddiagnose()
+{
+	$diag = $_POST['add'];
+	$id = $_GET['id'];
+	$rep = $_POST['diagnr'];
+	global $pdo;
+	$stmt = $pdo->query("SELECT DIAG FROM cars WHERE ID='$id'");
+	foreach($stmt as $row)
+	{
+		$diag1 = $row['DIAG'];
+	}
+	$stmt = $pdo->query("UPDATE cars SET DIAG = '$diag1 $diag', REP = '$rep' WHERE ID = '$id'");
+	$stmt->closeCursor();
+	$stmt=null;
 }
 			
 ?>
